Restrict OpenVPN/WireGuard imports
