+++ PassepartoutLibrary/Sources/PassepartoutVPN/Domain/Profile+OpenVPNSettings.swift
@@ -43,4 +43,20 @@ extension Profile {
             self.configuration = configuration
         }
     }
+
+    init(_ id: UUID = UUID(), name: String, configuration: OpenVPN.Configuration) {
+        let header = Header(
+            uuid: id,
+            name: name,
+            providerName: nil
+        )
+        self.init(header, configuration: configuration)
+    }
+
+    public init(_ header: Header, configuration: OpenVPN.Configuration) {
+        self.header = header
+        currentVPNProtocol = .openVPN
+        host = Host()
+        host?.ovpnSettings = OpenVPNSettings(configuration: configuration)
+    }
 }
